+++ CHANGES.md
@@ -4,7 +4,7 @@
 
 ### Improvements
 
--   Return JSON fields as dicts in read_dataframe (#556)
+-   Return JSON fields as dicts in `read_dataframe` (#556)
 
 ## 0.11.1 (2025-08-02)